Move facebook-auth to folder
