+++ backend/azure-functions/src/Auth/token-cache.ts
